dags/requirements.txt
ipykernel==4.10.1
appdirs==1.4.3
future==0.15.2
numpy==1.12.1
packaging==16.8
pandas==0.19.2
paramiko==2.2.1
py4j==0.10.4
pymssql==2.1.3
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.9
marshmallow==2.13.5
neo4j-driver==1.3.1
teamcity-messages==1.21
pylint==1.6.5
pytest==3.0.7
requests-mock==1.3.0
croniter==0.3.16
scp==0.10.2
neo4jupyter==0.1.2
py2neo==3.1.2
pyspark==2.1.2
findspark==1.3.0
